+++ torch/csrc/autograd/saved_variable.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <Python.h>
 #include <mutex>
 #include <memory>
 #include <functional>
