+++ torch/csrc/autograd/symbolic.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <Python.h>
 #include "torch/csrc/jit/ir.h"
 #include "torch/csrc/onnx/onnx.h"
 #include <vector>
