+++ torch/csrc/utils/python_stub.h
@@ -1,0 +1,4 @@
+#pragma once
+
+struct _object;
+typedef _object PyObject;
